Fix test discovery warnings about missing Vintage Story libraries
The mstest discovery was previously unable to find the VintageStory libraries. The AddVintageStoryPath module initializer adds the VintageStory library path to the assembly resolver path, but it only runs after class discovery.

So the path is now additionally passed as an assembly resolution path in the test.runsettings. In order to pick up those changes, the dotnet-test-explorer.testArguments and dotnet.unitTests.runSettingsPath settings now point directly to that runsettings file. However, those settings do not support variables, so they must point to the file in the source directory rather than the output directory.

The build previously copied the runsettings file to the output directory and modified a test parameter in it to point to the resources path in the source directory. Now that the runsettings is used directly from the source directory, that can no longer be done cleanly. So instead the test writes a new resourcespath.txt file in the output directory that has the absolute path of the resources folder in the source directory. The test code reads that file to get the path.